+++ artifactory/commands/python/poetry_test.go
@@ -1,16 +1,11 @@
 package python
 
 import (
-	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
-	"github.com/stretchr/testify/require"
-	"path/filepath"
-	"strings"
-	"testing"
-
 	"github.com/jfrog/jfrog-cli-core/v2/utils/tests"
 	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
 	"github.com/stretchr/testify/assert"
+	"path/filepath"
+	"testing"
 )
 
 func TestAddRepoToPyprojectFile(t *testing.T) {
@@ -34,28 +29,4 @@ func initPoetryTest(t *testing.T) (string, func()) {
 	assert.NoError(t, err)
 	poetryProjectPath, cleanUp := tests.CreateTestWorkspace(t, testAbs)
 	return poetryProjectPath, cleanUp
-}
-
-func TestGetPypiRepoUrlWithCredentials(t *testing.T) {
-	tests := []struct {
-		name        string
-		curationCmd bool
-	}{
-		{
-			name:        "test curation command true",
-			curationCmd: true,
-		},
-		{
-			name:        "test curation command false",
-			curationCmd: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			url, _, _, err := GetPypiRepoUrlWithCredentials(&config.ServerDetails{}, "test", tt.curationCmd)
-			require.NoError(t, err)
-			assert.Equal(t, tt.curationCmd, strings.Contains(url.Path, coreutils.CurationPassThroughApi))
-		})
-	}
 }
