+++ common/cliutils/utils.go
@@ -218,7 +218,6 @@ func CreateServerDetailsWithConfigOffer(createServerDetails func() (*config.Serv
 		return nil, err
 	}
 	log.Debug(fmt.Sprintf("Using <%s> server-id configuration", confDetails.ServerId))
-
 	// Take insecureTls value from options since it is not saved in config.
 	confDetails.InsecureTls = details.InsecureTls
 	confDetails.Url = clientUtils.AddTrailingSlashIfNeeded(confDetails.Url)
